fix: address CodeRabbit review round 7
- transport.hpp: 6-param create_usb_transport usb_pid uint16_t → int32_t (match sentinel)
- tactile_parser.hpp: resync after validation failure instead of dropping entire frame
- CI: add wujihandpy-tactile-bridge --help smoke test
- integration_demo: remove unused threading import, for i → for _
- tests: fix control_mode docstring int32 → uint16

+++ wujihandcpp/src/transport/usb.cpp
@@ -497,7 +497,7 @@ std::unique_ptr<ITransport>
 }
 
 std::unique_ptr<ITransport>
-    create_usb_transport(uint16_t usb_vid, uint16_t usb_pid, const char* serial_number,
+    create_usb_transport(uint16_t usb_vid, int32_t usb_pid, const char* serial_number,
                          int interface_num, unsigned char in_ep, unsigned char out_ep) {
     return std::make_unique<Usb>(usb_vid, usb_pid, serial_number, interface_num, in_ep, out_ep);
 }
